Only even numbers array iterator v1.1

+++ src/test/java/ru/job4j/iterator/EvenItTest.java
@@ -12,7 +12,7 @@ public class EvenItTest {
 
     @Before
     public void setUp() {
-        it = new EvenIt(new int[] {1, 2, 3, 4, 5, 6, 7});
+        it = new EvenIt(new double[] {1, 2, 3, 4, 5, 6, 7});
     }
 
     @Test(expected = NoSuchElementException.class)
@@ -38,13 +38,13 @@ public class EvenItTest {
 
     @Test
     public void  shouldReturnFalseIfNoAnyEvenNumbers() {
-        it = new EvenIt(new int[]{1});
+        it = new EvenIt(new double[]{1});
         assertThat(it.hasNext(), is(false));
     }
 
     @Test
     public void allNumbersAreEven() {
-        it = new EvenIt(new int[] {2, 4, 6, 8});
+        it = new EvenIt(new double[] {2, 4, 6, 8});
         assertThat(it.hasNext(), is(true));
         assertThat(it.next(), is(2));
         assertThat(it.hasNext(), is(true));
